Prevent deploying no changes
