+++ src/commands/revisionDraft/deployRevisionDraft/IDeployRevisionDraftContext.ts
@@ -8,6 +8,5 @@ import type { ExecuteActivityContext } from "@microsoft/vscode-azext-utils";
 import type { IContainerAppContext } from "../../IContainerAppContext";
 
 export interface IDeployRevisionDraftContext extends IContainerAppContext, ExecuteActivityContext {
-    baseRevisionName: string;
     template: Template;
 }
